Introduce cache layer
If multiple consecutive modifications happen to lie within the same
piece perform the operations "in place". In particular no new pieces
will be allocated if the changes occur at the end of the most recently
modified piece. In this case the piece is simply extended.

However changes in the middle of a piece involve memove(3) calls
which might hurt performance.

Since no new pieces are created the changes can't be undone on an
individual basis. The frontend should therefore call 'editor_snapshot'
at appropriate times inorder to invalidate the cache.