Make window_selection_get to return the correct range
This makes the operator operate on the correct file range if executed
from visual mode.

The problem is that if the cursor is visible the selection seems larger
than it actually is. The cell under the cursor is actually not part
of the selection eventhough it is visually indistinguishable from it.

The somewhat hacky way around this is to hide the cursor once selection
is active and show it again once the selection is cleared.

This will probably cause a headache if the cursor needs to be visible
while a selection is active like for example in the command prompt.