Make byte based iterator API handle the whole range [0, size]
Before this commit the valid range was [0, size) which represents
the file content. For the position at EOF (=size) a NUL byte is
returned which is not actually part of the underlying file.

This should fix various movements / editing operations at the end
of the file. For example Ctrl+w at the end of the command prompt.