Crit-bit tree based map
It supports lookups based on unique prefixes as well as ordered
iteration. More information about the data structure can be found
at:

 http://cr.yp.to/critbit.htm
 http://github.com/agl/critbit
 http://ccodearchive.net/info/strmap.html

It will be used to implement ":"-commands which means any unique
prefix will be recognized as a command.

It could also be used to implement (runtime) key bindings.