Preliminary user interface separation
In theory only ui-curses.[hc] should depend on curses, however in
practice keyboard input is still handled in vis.c. Furthermore the
syntax definitions as well as keyboard bindings and selection code
in window.c still depends on some curses constants.

There is also a slight regression in that the window status bar
does not show the current mode name. This and related global state
should be eliminated in the future.