text: safer temporary file creation
Set umask before calling mkstemp. According to POSIX 2008 this is
not necessary since the temporary file is guaranteed to be created
with permission restricted to the current user. However this is
more secure on non-conforming systems and safe as long as we do not
use multiple threads.

Fixes Coverity CID 101333.