vis: add a file open dialog
For this to work make sure you have vis-open and slmenu or dmenu
somewhere in $PATH.

For now the file dialog is shown for :open, :split and :vsplit
when the argument is either . (a dot) or looks like a file pattern.

For example

 :open *.[ch]

will show a listing of all C source and header files in the current
directory. Use a fuzzy search to make your choice.