vis: change key binding function prototypes
The idea is to work more like a finite state machine. Every
function gets an additional argument keys which holds the
already read keyboard input. The return value of the functions
should point to the first not consumed key. A return value
of NULL indicates that more input is needed. The function will
be called again from the editor core when more input is available.

These changes are mostly mechanical and in many cases not
optimal, they will be cleaned up in further commits.