text: get modification time after replacing file content in place
At the start of text_save_range we stat(2) the file to check whether
we have currently mmap(2)-ed it. Then we proceed to write the new
file content which changes modification time. Hence we have to
stat(2) again to retrieve it.

This should fix spurious warnings about file changes outside the
editor when editing e.g. symlinked files.