vis: improve switching to prompt mode
A call to vis_prompt_show will now automatically switch to prompt
mode. Within the prompt leave/enter handlers the focused window
(vis->win) will still point to the document window not the one
referring to the prompt.

The selection marks '< and '> are now only updated when a visual
mode is left.