vis-lua: implement vis.lines[..] array
This allows access to specific lines of a file, array indicies/line
numbers are 1 based. Read access is supported for [1, #lines] while
assigning to the 0 element will insert a new line at the beginning
of the file. Similarly assigning to lines[#lines+1] will add a new
line at the end of the file. The returned lines will not contain
any line termination characters.