text: introduce text_bytes_alloc0 utility function
Heap allocates a zero terminated string of the given range.
Freeing is the caller's responsibility.

Checks for unsigned integer overflow i.e. passing SIZE_MAX
as len will always fail because there is no room for the
terminating NUL byte. This is important as EPOS is defined
to be SIZE_MAX.