text-motion: add functions to get/set position based on display width
This is inherently a tricky thing to do because we cannot rely on
the current display state. The reason being that the position/cursor
which is modified might not currently be in the visible area.

Tabs are a particular problem because they have a variable display
width. However this new code is certainly not worse than the current
behaviour which relies on text_line_char_{get,set} and thus simply
counts graphemes.

Not yet completely convinced that this is the right approach.