view: maintain ordering relation among cursors
This is currently only enforced upon creation i.e. we assume
that after creation a cursor can not change its relative ordering
with respect to its neighbors.

The existing code assumes that when iterating through cursors with:

 for (Cursor *c = view_cursors(view); c; c = view_cursors_next(c))
    ...

new cursors created with view_cursors_new do not show up.
This assumption is preserved under the following conditions:

 * it only holds for the most recent view_cursors call

   As a consequence when doing nested iterations new cursors
   will be yielded once the inner view_cursors call was performed.

 * view_cursors_primary_get is not called