view: change internal representation of selections
We place the end mark inside the selection as opposted to on the character
immediately following it. This is better when selections are touching each
other. Previously for two seletions [a][b] the end mark for selection
a would be at the same location as the start mark of selection b. Thus
when for example the content of selection b is deleted it would also
destroy selection a, because the end mark would no longer be valid.