vis: make af and if text objects more robust
It still assumes that a function is terminated by the "\n}" seqeuence
and uses the same code as the % motion for matching curly braces.
Hence this won't work for functions containing an unbalanced
number of { or } symbols (e.g. due to comments or strings etc).

Also due to the way text objects are implemented in general the inner
variant will only work properly if the cursor is located somewhere
within the function. Otherwise it will include everything from the
current cursor position up to the end of the function body.