statusline: add total line number
This adds a helper function to get the total number of lines in a file (which
would be expensive for large files). As well as adding that information to the
default status line for a window. Since the statusline information is protected
by the `UI_OPTION_LARGE_FILE` check, this code is not run for large files
anyway, so calculating the total line number should not be too costly.

It's possible, however, that the line count should be done elsewhere so that it
is not repeated for every statusline update.