statusline: add percentage position within file
With the previous calculation of the number of lines in the file, this is
trivial, and helps with keeping track of where in the file the cursor is. It's
also what vim does by default, though vim uses `TOP` and `BOT` and `ALL` when
the relevent parts of the file are displayed. I'm not sure these additions add
much.