sam: improve ^ matching start of line
There are some nasty differences between the meaning of ^ in
Plan 9's regexp library and POSIX when using REG_NEWLINE.
The former only matches at the beginning of a line wheras the
latter matches the zero-length string immediately after a
newline character \n. As a result this also matches after the
very last newline at the end of the file.

This is undesired behavior for a command like :x/^/c/#/
Hence we try to filter out this last match.

Close #264