vis: do not run registered event handlers for internal files
This fixes interactive :-commands when the user has configured
to set custom options vis:command(...) via the Lua win_open
event handler.

The problem was that the creation of the window for the command
prompt would itself trigger an execution of a :-command. Upon
successful completion the editor would switch to normal mode.
Therefore the interactively entered command would not be applied
to the correct range.