vis: make j and k a linewise inclusive motion
They behave like an inclusive motion, but only if they are also
linewise (which they are by default).

This should make `yjp` and `ykp` yank both the current and
the next/previous line when the cursor is at the start of
a line.

See also 532f52e9e52b98dc5749396f7353295418e0227a and #237