vis: introduce vis-menu helper script
This serves as a wrapper around dmenu(1) and slmenu(1), by default
the latter is preferred because it also works without an X server.

The program and its default arguments can be configured by means of
the $VIS_MENU and $VIS_MENU_ARGS environment variables.

Eventually we might decide to inlcude a stripped down version of
slmenu in the vis source tree.