vis: tweak completion commands
Change tr command to split words, this won't properly work with
Unicode but should at least avoid unwanted non-word symbols and
be POSIX conformant. A possible alternative would be to use
grep -o -E '\w+' while the -o option is not part of POSIX it
seems to be mostly supported. However the \w regex syntax might
not be supported.

Force ls(1) output to be linewise and surpress the same prefix.

Change sed invocation to only replace proper prefixes.