vis: fix inner text object variants when cursor is on opening delimiter
Except for special cases like gn and gN vis expected that a text object
would be a function mapping a position to a range as follows:

  f: pos -> [start, end]      with     start <= f(pos) <= end

Clearly this condition does not hold for inner text objects when the
initial position i.e. the cursor is on the opening delimiter.

This also obsoletes the need for the SPLIT text object flag which should
be removed in a later commit if the current behavior is found to be
working as expected.