vis: improve unpaired text objects
For unpaired text objects delimited by ", ' or ` if there is no preceding
symbol on the same line, advance starting position to first occurence on
the same line.

As a result ci" can be used to change the inner quotes on the same line
even if the cursor is currently to the left of the opening quote.

If the line contains no such symbol at all then the text objects will
will continue to match across line boundaries. This behavior is different
in vim where for example a ci" command on a line without any quotes has
no effect.

Close #358