sam: change license header
We use an adapted variant of sam's structural regular expression
based command language. The initial implementation was partially
based upon the following functions from sam / acme:

 * parse.h / edit.h (struct definitions)
 * cmd.c / edit.c (functions parsecmd, simpleaddr, compoundaddr)
 * xec.c / ecmd.c (cmdexec)
 * address.c / addr.c (address)

It turns out the relevant code can be traced back to the initial
X11 port of sam which is distributed under an ISC-like license
instead of the Lucent Public License Version 1.02 used for Plan 9,
plan9port and 9base.

 http://www.netlib.org/research/
 http://www.netlib.org/research/sam.shar

Hence we switch to the simpler license variant.

Close #238