vis: improve `r` in normal and replace mode
In normal mode `r<key>` was previously implemented as `R<key><Escape>`.
However this does not work when the replacement key is `<Enter>` to insert
a new line, because in replace mode new lines are not overwritten.

The count is now also respected.

Also properly support `r` in visual mode where before it was aliased to `c`.

Fix #190