vis: treat vis_keys_feed input like a macro replay
We need to push keys individually to the input queue such that
the state machine can advance and record keys into the operator
macro if necessary.

Previously feeding the following input:

 isome text<Escape>.

would not work as expected because the complete key stream
was pushed to the input queue at the same time during which
the operator macro was not yet active. Thus the dot command
at the end would have nothing to repeat.