vis-lua: introduce pre-save hook
The first argument is the file object while the second argument denotes
the full path to which it will be written. Path might be `nil` if the
file is going to be written to stdout.

The Lua function is expected to return a boolean value indicating whether
the write operation should proceed or be aborted.