vis: add `:set escdelay nn` option
Make the delay used to distinguish between an <Escape> key and
other terminal escape sequences such as for the Meta key run
time configurable. The value is given in miliseconds and defaults
to 50ms.

Notice that terminal multiplexers like dvtm or tmux might also
induce some delay which has to be configured independently.