vis: re-open /dev/tty read-writeable
libtermkey's initialization routine tries to write to the underlying
file descriptor which fails with EBADF if it is opened read only.
This was a problem in the terminal restore code called after a shell
command is executed.

It should fix the following:

 $ echo foo | vis - > bar
 :!/bin/sh
 exit

where before vis would no longer accept any input.