vis: redirect stdout and stderr streams to /dev/null when lacking a consumer
If the caller of vis_pipe is not interested in the output, redirect it
to /dev/null and close the pipe. Otherwise we would wait for possible
output (which might never arrive) only to throw it away.

As a consequence background processes can now be started with:

 :> { plumber <&3 3<&- & } 3<&0 2>&-

whereas before one also had to explicitly close stdout:

 :> { plumber <&3 3<&- & } 3<&0 1>&- 2>&-