vis: avoid inconsistent state after file reloading with :e
When a file is opened in multiple windows we could end up in an
inconsistent state by doing:

 $ vis some-file
 :spl
 :e

At this point the two windows were referencing different file objects
thus changes in one would not affect the other although they have
the same path.

Another option would be to reload all windows currently displaying
the file being reloaded. We opt for the simpler fix for now.