vis: use file(1) for file type detection
File type detection works as follows:

 1) strip off suffixes to be ignored and test against a
    set of known file extensions

 2) run `file -bL --mime-type` and check against a set
    of known mime types

 3) read out the first few bytes of the file and pass
    them to custom Lua file type detection functions

For now the configured file extensions are literal strings
which are matched against the end of the file name. Maybe
we should use Lua patterns instead.

We will need to add more mime types to our mapping table.
For now only the `bash` file type was associated with the
text/x-shellscript mime type.