vis-lua: add preliminary version of LDoc based Lua API documentation
A new Makefile target `luadoc` has been added which generates
HTML documentation in the doc subfolder using ldoc(1) from
https://stevedonovan.github.io/ldoc/

There are still a few problems to resovle, for example the links (e.g.
in parameter lists) to our custom types seem to be broken.

At this point only the C part of the Lua API is covered and even
that is not yet complete.