Move all lua related files to lua/ subfolder
Also remove the lexers sub directory from the Lua search path.
As a result we attempt to open fewer files during startup:

 $ strace -e open -o log ./vis +q config.h && wc -l log

In order to avoid having to modifiy all lexers which `require('lexer')`
we instead place a symlink in the top level directory.

 $ ./configure --disable-lua
 $ rm -rf lua

Should result in a source tree with most lua specifc functionality
removed.