build: improve description of --enable-lpeg configure option
It links lpeg statically into the vis binary, meaning it does not need
to be dlopen(3)-ed at runtime. While this improves the portability of
the resulting vis binary, it is not necessary to enable this option to
get syntax highlighting support.

Maybe the option should be renamed to avoid further confusion.