vis-lua: implement file.marks[] array
We need to retrieve a pointer to the File struct but can not store it
directly in the Lua uservalue because the address is already used for
the regular file object (of type vis.file). For now we use file->marks
as an address and then use offsetof to retrieve the start of the struct.