view: add infrastructure for delayed cursor destruction
At least one cursor (referred to as primary or main cursor) has always
to exist. In the sam command language implementation we might want to
dispose a cursor even if it is the primary one before later commands
will create different ones (e.g. `:x/pattern/ { i/>>>/ a/<<</ }`).

This commit introduces view_cursors_dispose_force. If called on
the last remaining cursor, its selection is cleared and it is marked for
destruction as soon as a new cursor is created. view_cursor_disposed
returns the cursor marked for deletion (if any) and clears the descruction
flag.