sam: remove s (substitute) command use x (extract) and c (change) instead
They are roughly equivalent, instead of

 s/pattern/replacement/

to replace the first occurrence of pattern you can specify an address
to the change command:

 /pattern/ c/replacement/

the only difference being that the first command is restricted to the
current line.

 -+x/pattern/ c/replacement/

also restrictes matches to the current line, but performs the substitution
on the whole line not only the first match. Currently it is not possible
to only replace the n-th match as `s2/pattern/replacement/` would do
in sam(1).

A possible alternative syntax generalizing this concepts and applying
it to the `x` and `g` commands will be investigated in the future.

Global substitution as in

 %s/pattern/replacement/g

can be performed using

 x/pattern/ c/replacement/