view: keep track of the most recently created cursor
Previously this was done implicitly through the primary cursor
which was always adjusted when a new cursor is being created.

However, this is no longer the case and we do not want to iterate
through all cursors when creating a new one.

In the longterm we might want to store cursors in a contiguous
memory location (i.e. an array) instead of chaising pointers
all over the place.