travis: try to fix code coverage for test/core
The test/core tests use source files from the repository root
(e.g. ../../text.c) which seems to confuse gcov and/or codecov.io
because the same file is used under multiple paths.

We add a new entry to the test matrix which just runs the core/tests.
This is done by invoking the `make clean` target and skipping any
tests which require the main vis binary.