appveyor: fix symlinks for Cygwin builds
Git for windows is used to checkout the repository and it
creates broken symlinks (files containing the target) which
break the tests because e.g. the lexer.lua module can not
be loaded.

We use a bash script to convert these pseudo symlinks:
http://stackoverflow.com/questions/38140373
and invoke it via powershell because that seems to be the
only way to execute multiline commands from within the
AppVeyor YAML configuration file.