Add experimental raw vt100 UI backend
The intention of this is not to slowly reimplement curses but to provide
a minimal working terminal UI backend which can also be used for debugging,
fuzzing and in environments where curses is not available.

Currently no attempt is made to optimize terminal output. The amount of
flickering will depend on the smartness of your terminal emulator.