Fix errors and add ANS Forth 2012 keywords
Added all ANS Forth 2012 keywords as defined at
http://lars.nocrew.org/forth2012/core.html and removed keywords
that were not part of the standard. This necessitated rewriting
most of the Strings rules as well as removing some rules not
consistent with the standard. Only the s\" form should allow
escaping. The list of characters which may appear as part of a
keyword has also been expanded where appropriate.

Because '.' is a keyword as well as the first chatacter in a string
pattern, strings must now be given parsing precedence over keywords
to ensure proper highlighting.

A few errors were also fixed such as moving the true (which should
make keywords case-insensitive) within the word_match function's
closing paren. Parens have been removed from the operator list and
moved to their correct place as the delimiters for block comments.