Clean the window opening and focusing process.
Until now, vis was focusing on a window before it was open.
Now the process is:
- A window is opened, and the `WIN_OPEN` event is fired.
- Some more configuration of the window may happen.
- Vis focuses on the window.

Note that it implies that when `WIN_OPEN` is fired, the active window is *not*
the newly created window. So lua code suscribing to the event may want to
specify the window as a parameter to `vis:command`.

    vis.events.subscribe(vis.events.WIN_OPEN, function(new_win)
      -- This command applies to the currently active window, *not* to the newly
      -- created window.
      vis:command('set numbers')
      -- This command applies to the newly created window.
      vis:command('set numbers', new_win)
    end)