vis: validate window and file objects after event handlers are run
If the Lua event handler does something stupid it can invalidate the
objects we are dealing with. As an example if a WIN_OPEN event handler
contains a line like:

 vis:command('q!')

to immediately close the window this causes havoc. The reason being that
the window resources will be free(3)-ed while the C code calling the event
handler still has references to it. For now this is deemed as an invalid
event handler, but at least we should no longer segfault.