sam: use default address for file looping commands
This sould fix a clang analyzer warning and make the code more robust.
The problem could actually not occur in practice because the X and Y
commands always have the internal cmd_select as intermediate node in
the execution tree.

In general the multiple file support will need more design work.

Fix #551