vis: ensure complete ! command output is displayed
For interactive processes started using `:!` stdout is redirected
to stderr normally used by vis to draw its user interface.
For some reason the first byte written by the interactive application
is not being displayed. I suspect it has something to do with the
terminal state change. For now we are writing a dummy space (which
is never shown) ourself to ensure that the complete output is visible.

Fix #545