Fix paragraph object when cursor in a empty line
When the cursor is in an empty line, bothe the previous and next
paragraph was affected by the paragraph text objects (dip, vap...).
This was due to vis looking for the closest start and stop of
paragraph at the cursor position.

This patch make vis first look for the next end of paragraph, and look
for the start of the paragraph upward from there, hence selecting the
/next/ paragraph when the cursor is between two paragraphs on a blank
line.