vis: process aliased key sequences individually
While the complete alias is added to the input queue,
the called key bindings should only see the keys they
have asked for.

Previously a mapping such as:

 :map! normal gv \"^Sv

did not work as expected because the key binding for
the register did reject an invalid multi letter register
name.

Might also influence #581