vis-complete: Slight refactoring
Move the "case $PATTERN" block inside "if $COMPLETE_WORD" to make it
clear that it is specific only to the "else" branch.

Eliminate the $START variable - it was used only once, and using
dirname(1) directly is obvious enough.

Remove the comments inside the "case" block (explaining the "what") and
replaced them with a single comment explaining the "why".