sam: fix g/^$/
With POSIX ERE the pattern ^$ matches strings ending with a new
line because an empty match is reported after the trailing newline
at the very end of the input.

This is undesirable for use cases like

  x g/^$/ d

which is supposed to delete all empty lines of a file.

As a fix we disregard empty matches at the end of the given range.