Added OS check in target 'vis-single-payload.inc'
Reading through the Github issues,it sounds as if this was meant for
Linux and the if statement reflects that. If that's not the case it can
be changed.
( https://github.com/martanne/vis/issues/162 )

FreeBSD's `od` has a different format compared to the GNU Coreutils version.
The output format is also a little bit different. For some reason both
`sed` and `gsed` on FreeBSD can't find the end of file if commands have
been previously passed with `-e` so I pipe from sed into sed again.

Added `-f -` to both tar commands to make it obvious as to where the output
is going.