ui: make sure $TERM is set
libtermkey passes the $TERM value unchecked to libunibilium which just
aborts (in debug builds) or crashes (in release builds). Workaround
that by defaulting to `xterm`, if $TERM is unset. It should eventually
be fixed in libtermkey/unibilium too.

This fixes test suite failures on the Debian package build environment.
It might also be the reason for failures in other CI environments
e.g.: https://github.com/alpinelinux/aports/pull/3768