vis: Fix a crash when re-entering prompt window
vis->mode was incorrectly "restored" to a NULL prompt->parent_mode in
prompt_restore() because vis_prompt_show() was assuming that the
currently active window is always a regular one.
But if the active window is the command window, a new command window(s) would
be created with the old command window as its parent.