Add ability to access `tabwidth` from Lua
This commit adds the ability to access the current tabwidth from a Lua
plugin. This can be used, for instance, to calculate the current column
in terms of character cells, instead of Unicode codepoints. I.e. a TAB
character could be counted as `tabwidth` character cells, instead of as
one.