vis: fix search wrapping bugs
1) “$” matches in the middle of the text.
visvis
   ^   - standing here
    \/ - at first we search forward-\
\_/    - wrap, if nothing found <---/

After wrapping, in the second range “$” will treat end of the range
as EOL so “/vis$” will wisely match and moves cursor to the first
column.

2) No match after wrapping.
vissssss
 ^^       - standing here or here
 \\____/  - search this before wrapping ---\
V         - search range after wrapping <--/

“/vis” will *not* match (after wrapping), because it crosses ranges.

---

So the real solution would be that instead of the end position, the
start position of the possible match should be limited because a match
can cross the search ranges. To keep things simple, simply search two
whole text after wrapping.

visvis
\____/