text: ignore fsync(2) errors on unsupported directory descriptors
When saving a file by atomically renaming it to its final destination,
we fsync(2) the parent directory to make sure the new directory entry
is persisted.  However, not all file systems support fsync on file
descriptors referring to directories. As a result the save operation
fails and subsequent attempts result in warnings regarding outdated file
content, even though the data has most likely been successfully written.

Ignoring this particular error seems fine, because it is a permanent
limitation of the file system and not a temporary failure.

Fixes #792