color-column: Don't change fg/bg if not set explicitly
eg. if your long line is a comment with green fg, and you set your
column color bg red while not specifying the fg, then the result is
green fg on red bg.

Prior to this change the result would be default fg on red bg, thus
one char in the long line of green text would look odd/wrong.

Of course if you do explicitly set the column color fg to default in your
theme then the result will not be what you expect - ideally we need
an UNSPECIFIED color type instead of relying on DEFAULT.