vis-open: allow to show files vertically
It might be more comfortable to navigate through the files
arranged vertically.

Add ability to specify VIS_OPEN_LINES environment variable
which is passed to vis-menu as -l option.

It would be better to add vis option for this to set it
via lua config, but it might be added later.

Signed-off-by: Vadym Kochan <vadim4j@gmail.com>