vis-complete: General changes to `find` behaviour
Ignore hidden paths only if pattern's `basename` does not start with ".";

Once we are limiting `find` output with `head`, it makes sense to also
use the -maxdepth option;

Unquote filename prefixes. Filename completion it's mostly done with
usual characters. Even if the filename STARTS with a meta-character,
we can always use an empty prefix and type the prefix inside vis-menu.