vis-complete: Allow to complete text objects other than short words
Without "--word" and "--file" options, vis-complete now reads the
standard input. Since we are enabling a more flexible use, it makes
sense to provide options to display vis-menu in multiple lines [-l]
and to use a prompt [-p]. This is useful to complete lines or long
words in an external file, for example.